Completing implementation for Simulator and Agent

+++ Order_structures/order.cpp
@@ -3,7 +3,7 @@
 Order::Order(int price, int id, int trade_type)
     : price(price), id(id), trade_type(trade_type) {}
 
-int Order::getPrice() const { return price; }
+float Order::getPrice() const { return price; }
 
 int Order::getTradeType() const { return trade_type; }
