+++ Trade_logic/trades.cpp
@@ -3,24 +3,24 @@
 #include <iostream>
 #include <fstream>
 
-Order set_order(int price, int id, int type) {
+Order set_order(float price, int id, int type) {
     Order order(price, id, type);
     return order;
 }
 
-int calculate_spread(Order& buy_order, Order& sell_order) {
+float calculate_spread(Order& buy_order, Order& sell_order) {
     return sell_order.getPrice() - buy_order.getPrice();
 }
 
 // Matches a single incoming order against the opposite book
 bool match_orders(Order& order, Orderbook& buy, Orderbook& sell) {
     int trade_type = order.getTradeType();
-    int price = order.getPrice(); 
+    float price = order.getPrice(); 
     bool matched = false;
     
     if(trade_type == 0){
         while(!(sell.is_empty())){
-            int lowest_sell_price = sell.find_lowest_price();
+            float lowest_sell_price = sell.find_lowest_price();
             if(lowest_sell_price == -1){
                 return false;
             }
@@ -29,7 +29,7 @@ bool match_orders(Order& order, Orderbook& buy, Orderbook& sell) {
                 if (!sell_order.has_value()){
                     return false;
                 }
-                int spread = calculate_spread(order, *sell_order);
+                float spread = calculate_spread(order, *sell_order);
                 log_trade(order, *sell_order, lowest_sell_price, spread);
                 sell.removeOrder(*sell_order);
                 matched = true;
@@ -42,7 +42,7 @@ bool match_orders(Order& order, Orderbook& buy, Orderbook& sell) {
 
     if(trade_type == 1){
         while(!(buy.is_empty())){
-            int highest_buy_price = buy.find_highest_price();
+            float highest_buy_price = buy.find_highest_price();
             if(highest_buy_price == -1){
                 return false;
             }
@@ -51,7 +51,7 @@ bool match_orders(Order& order, Orderbook& buy, Orderbook& sell) {
                 if (!buy_order.has_value()){
                     return false;
                 }
-                int spread = calculate_spread(*buy_order, order);
+                float spread = calculate_spread(*buy_order, order);
                 log_trade(*buy_order, order, highest_buy_price, spread);
                 buy.removeOrder(*buy_order);
                 matched = true;
@@ -67,8 +67,8 @@ bool match_orders(Order& order, Orderbook& buy, Orderbook& sell) {
 // Sweeps resting orders in both books and matches them against each other
 void sweep_book(Orderbook& buy, Orderbook& sell) {
     while (!buy.is_empty() && !sell.is_empty()) {
-        int highest_buy_price = buy.find_highest_price();
-        int lowest_sell_price = sell.find_lowest_price();
+        float highest_buy_price = buy.find_highest_price();
+        float lowest_sell_price = sell.find_lowest_price();
 
         if (highest_buy_price == -1 || lowest_sell_price == -1) break;
 
@@ -78,7 +78,7 @@ void sweep_book(Orderbook& buy, Orderbook& sell) {
 
             if (!buy_order.has_value() || !sell_order.has_value()) break;
 
-            int spread = calculate_spread(*buy_order, *sell_order);
+            float spread = calculate_spread(*buy_order, *sell_order);
             log_trade(*buy_order, *sell_order, lowest_sell_price, spread);
             buy.removeOrder(*buy_order);
             sell.removeOrder(*sell_order);
@@ -118,9 +118,9 @@ void prepare_file(void) {
     }
 }
 
-void log_trade(Order& buyer, Order& seller, int price, int spread) {
+void log_trade(Order& buyer, Order& seller, float price, float spread) {
     prepare_file();
-    //TODO: logic to update state of agents
+    //TODO: logic to update state of agents + update thier cash/shares based on the trade.
     std::ofstream log_file("trade_log.csv",      std::ios::app);
     if (log_file.is_open()) {
         log_file << buyer.getId()
