+++ Trade_logic/trades.h
@@ -1,7 +1,7 @@
-#include "orderbook.h"
+#include "../Order_structures/orderbook.h"
 
 // Returns an order object with the given price and id
-Order set_order(int price, int id, int type);
+Order set_order(float price, int id, int type);
 
 // Returns true if the order matches with an order in the orderbook, false
 // otherwise. Function is called immideately after a new order is called Takes
@@ -14,7 +14,7 @@ bool sell_trade(Order &order, Orderbook &buy, Orderbook &sell);
 bool match_orders(Order &order, Orderbook &buy, Orderbook &sell);
 void sweep_book(Orderbook& buy, Orderbook& sell);
 
-void log_trade(Order &buyer, Order &seller, int price, int spread);
-int calculate_spread(Order &buy_order, Order &sell_order);
+void log_trade(Order &buyer, Order &seller, float price, float spread);
+float calculate_spread(Order &buy_order, Order &sell_order);
 
 void prepare_file(void);
